requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
beautifulsoup4==4.7.1
bs4==0.0.1
cycler==0.10.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
hdf5storage==0.1.15
isort==4.3.20
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.4
Markdown==3.1.1
matplotlib==3.1.0
mccabe==0.6.1
mock==3.0.5
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==6.0.0
protobuf==3.7.1
pylint==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
PyYAML==5.1
scipy==1.3.0
six==1.12.0
soupsieve==1.9.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
typed-ast==1.3.5
Werkzeug==0.15.4
wrapt==1.11.1
